refactor: update imports and add new model classes for job and trial results
