+++ rock/sdk/agent/models/trial/__init__.py
@@ -1,3 +1,17 @@
 from .config import AgentConfig, ArtifactConfig, EnvironmentConfig, TaskConfig, VerifierConfig
+from .result import AgentInfo, AgentResult, ExceptionInfo, ModelInfo, TimingInfo, TrialResult, VerifierResult
 
-__all__ = ["AgentConfig", "EnvironmentConfig", "VerifierConfig", "TaskConfig", "ArtifactConfig"]
+__all__ = [
+    "AgentConfig",
+    "EnvironmentConfig",
+    "VerifierConfig",
+    "TaskConfig",
+    "ArtifactConfig",
+    "TrialResult",
+    "AgentInfo",
+    "ModelInfo",
+    "AgentResult",
+    "VerifierResult",
+    "ExceptionInfo",
+    "TimingInfo",
+]
